Changed Database interfaces to use Long instead of Integer.

+++ server/src/main/java/com/example/demo/repository/CapacitorTypeRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.CapacitorType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CapacitorTypeRepository extends JpaRepository<CapacitorType, Integer> {
+public interface CapacitorTypeRepository extends JpaRepository<CapacitorType, Long> {
 }
